# [CWS] improve glob matching by pre-computing segments