--- a/pkg/security/secl/compiler/eval/pattern.go
+++ b/pkg/security/secl/compiler/eval/pattern.go
@@ -58,36 +58,64 @@ func hasPrefix(s, prefix string, caseInsensitive bool) bool {
 
 // PatternMatches matches a pattern against a string
 func PatternMatches(pattern string, str string, caseInsensitive bool) bool {
-	if pattern == "*" {
+	patternElem := newPatternElement(pattern)
+	return PatternMatchesWithSegments(patternElem, str, caseInsensitive)
+}
+
+// PatternMatchesWithSegments matches a pattern against a string
+func PatternMatchesWithSegments(patternElem patternElement, str string, caseInsensitive bool) bool {
+	if patternElem.pattern == "*" {
 		return true
 	}
 
-	if len(pattern) == 0 {
+	if len(patternElem.pattern) == 0 {
 		return len(str) == 0
 	}
 
-	for len(pattern) > 0 {
-		star, segment, nextIndex := nextSegment(pattern)
-		if star {
+	for _, seg := range patternElem.segments {
+		if seg.star {
 			// there is no pattern to match after the last star
-			if len(segment) == 0 {
+			if len(seg.segment) == 0 {
 				return true
 			}
 
-			index := index(str, segment, caseInsensitive)
+			index := index(str, seg.segment, caseInsensitive)
 			if index == -1 {
 				return false
 			}
-			str = str[index+len(segment):]
+			str = str[index+len(seg.segment):]
 		} else {
-			if !hasPrefix(str, segment, caseInsensitive) {
+			if !hasPrefix(str, seg.segment, caseInsensitive) {
 				return false
 			}
-			str = str[len(segment):]
+			str = str[len(seg.segment):]
 		}
-		pattern = pattern[nextIndex:]
 	}
 
 	// return false if there is still some str to match
 	return len(str) == 0
+}
+
+type patternElement struct {
+	pattern  string
+	segments []patternSegment
+}
+
+type patternSegment struct {
+	star    bool
+	segment string
+}
+
+func newPatternElement(element string) patternElement {
+	el := patternElement{
+		pattern: element,
+	}
+
+	for len(element) > 0 {
+		star, segment, nextIndex := nextSegment(element)
+		element = element[nextIndex:]
+		el.segments = append(el.segments, patternSegment{star, segment})
+	}
+
+	return el
 }