### What does this PR do?

This PR improves the performance of the glob matching, by keeping track of the segments instead of re-computing for each match operation.
The segment computation is also sped up by skipping the utf-8 decoding

benchmark:
```
# before
BenchmarkGlob-10    	15692864	        74.17 ns/op	       0 B/op	       0 allocs/op
# after
BenchmarkGlob-10    	22775433	        52.72 ns/op	       0 B/op	       0 allocs/op

# before
BenchmarkNextSegment-10    	124941762	         9.490 ns/op	       0 B/op	       0 allocs/op
# after
BenchmarkNextSegment-10    	178843194	         6.648 ns/op	       0 B/op	       0 allocs/op
```

### Motivation

<!--
* What inspired you to submit this pull request?
* Link any related GitHub issues or PRs here.
-->

### Additional Notes

<!--
* Anything else we should know when reviewing?
* Include benchmarking information here whenever possible.
* Include info about alternatives that were considered and why the proposed
  version was chosen.
-->

### Possible Drawbacks / Trade-offs

<!--
* What are the possible side-effects or negative impacts of the code change?
-->

### Describe how to test/QA your changes

<!--
* Write here in detail or link to detailed instructions on how this change can
  be tested/QAd/validated, including any environment setup.
-->